fix: resolve critical linting errors
- Fix errcheck: add error handling for all unchecked error returns
- Fix errorlint: use errors.Is instead of == for error comparison
- Fix gofmt: correct formatting issues in config and models
- Fix gosimple: remove unnecessary nil checks around range loops
- Fix govet: resolve variable shadowing in commands.go
- Fix revive: remove zero value from variable declaration
- Fix unconvert: remove unnecessary type conversions
- Rename DEFAULT_TIME to DefaultTime for proper Go naming
- Fix syntax errors in login/packets.go from nil check removal

+++ internal/channel/handler/npc.go
@@ -38,7 +38,7 @@ func (h *NPCSelectHandler) Handle(s game.Session, reader *packet.Reader) {
 		return
 	}
 
-	objectID := uint32(reader.ReadInt())
+	objectID := reader.ReadInt()
 	x := reader.ReadShort()
 	y := reader.ReadShort()
